Split RobloxPlatform into multiple files
