+++ CMakeLists.txt
@@ -24,7 +24,12 @@ target_sources(Luau.LanguageServer PRIVATE
         src/Utils.cpp
         src/CliConfigurationParser.cpp
         src/platform/LSPPlatform.cpp
-        src/platform/RobloxPlatform.cpp
+        src/platform/roblox/RobloxCodeAction.cpp
+        src/platform/roblox/RobloxColorProvider.cpp
+        src/platform/roblox/RobloxCompletion.cpp
+        src/platform/roblox/RobloxLuauExt.cpp
+        src/platform/roblox/RobloxSourcemap.cpp
+        src/platform/roblox/RobloxStudioPlugin.cpp
         src/operations/Diagnostics.cpp
         src/operations/Completion.cpp
         src/operations/DocumentSymbol.cpp
